revert some things to try and get this working
